+++ src/main/kotlin/com/github/shiraji/emoji/data/EmojiDataManager.kt
@@ -1,5 +1,5 @@
 package com.github.shiraji.emoji.data
 
 object EmojiDataManager {
-    lateinit var emojiList: List<EmojiData>
+    val emojiList: MutableList<EmojiData> = mutableListOf()
 }
